+++ R/getPlayerOrientation.R
@@ -27,13 +27,18 @@ getPlayerRotation <- function(player_id = NULL)
 {
     p <- NA
 
-  if(is.null(player_id)){
-    while(is.na(p)) p <- mc_sendreceive("player.getRotation()")
-  } else {
-    while(is.na(p)) p <- mc_sendreceive(merge_data("entity.getRotation", player_id))
-  }
+    if(is.null(player_id)){
+        # API seems to give wrong results for many angles!
+        # while(is.na(p)) p <- mc_sendreceive("player.getRotation()")
+        d <- getPlayerDirection(player_id)
+        p <- atan2(d[3], d[1])*180/pi
+        p <- (p - 90) # convert to minecraft rotation angle
+    } else {
+        while(is.na(p)) p <- mc_sendreceive(merge_data("entity.getRotation", player_id))
+    }
 
-  as.numeric(p)
+    p <- as.numeric(p)
+    (p + 360*10) %% 360 # convert to angle 0-360
 }